[lib] Move array() function from init.c to public librpminspect API
This is used both when initializing rpminspect when reading config
files, but also by the license inspection when reading the JSON
database with the license information.
